update touchbar play pause btn
